flask/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
eventlet==0.36.1
PyMySQL==1.1.1
qrcode[pil]==7.4.2
Werkzeug==3.0.3
